+++ src/api-hooks/InfoCategory.ts
@@ -1,19 +1,40 @@
 import { httpClient } from '.';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { ICategoryData, ICategoryQueryRequest, ICategoryResponse } from '../@types/InfoCategory';
+import { useEffect, useState } from 'react';
+import toast from 'react-hot-toast';
 
 export function useGetAllCategoryRequest() {
-  return useQuery([`/admin/article/category`], () =>
+  const [options, setOptions] = useState<ICategoryData[]>([]);
+
+  const { data, error } = useQuery([`/admin/article/category`], () =>
     httpClient<ICategoryData[]>({
       method: 'GET',
       url: `/admin/article/category`,
     }),
   );
+
+  useEffect(() => {
+    if (data) {
+      setOptions(data.data);
+    } else if (error) {
+      toast.error('카테고리 목록을 불러오는데 실패했습니다.');
+    }
+  }, [data, error]);
+
+  return { options, setOptions };
 }
 
 // 해당 호출을 적용한 컴포넌트 렌더링 시 바로 query가 실행되지 않고 트리거 발생시 refetch 등을 통해 실행되게 하려면 isEnabled 속성 추가하기
 export function useGetCategoryRequest(params: ICategoryQueryRequest, isEnabled?: boolean) {
-  return useQuery(
+  const [categoryList, setCategoryList] = useState<ICategoryData[]>([]);
+  const [totalPosts, setTotalPosts] = useState(1);
+
+  const {
+    data,
+    error,
+    refetch: getCategoryRefetch,
+  } = useQuery(
     [`/admin/article/category/page?page=${params.page}`, params],
     () =>
       httpClient<ICategoryResponse>({
@@ -22,6 +43,21 @@ export function useGetCategoryRequest(params: ICategoryQueryRequest, isEnabled?:
       }),
     { enabled: isEnabled },
   );
+
+  useEffect(() => {
+    if (data) {
+      setCategoryList(data?.data.articleTypeList);
+      setTotalPosts(data?.data.totalPage);
+    } else if (error) {
+      toast.error('카테고리 목록을 불러오는데 실패했습니다.');
+    }
+  }, [data, error]);
+
+  return {
+    categoryList,
+    totalPosts,
+    getCategoryRefetch,
+  };
 }
 
 export default function useCreateCategoryRequest() {
@@ -34,14 +70,41 @@ export default function useCreateCategoryRequest() {
   );
 }
 
-export function useUpdateCategoryRequest() {
-  return useMutation((data: ICategoryData) =>
+export function useUpdateCategoryRequest(
+  currentPage: number,
+  setCurrentPage: React.Dispatch<React.SetStateAction<number>>,
+  onChangeViewMode?: () => void,
+) {
+  const { getCategoryRefetch } = useGetCategoryRequest({
+    page: currentPage - 1,
+  });
+
+  const {
+    mutate: categoryUpdateRequest,
+    data,
+    error,
+  } = useMutation((data: ICategoryData) =>
     httpClient<ICategoryResponse>({
       method: 'PATCH',
       url: `/admin/article/category/${data.id ? data.id : ''}`,
       data,
     }),
   );
+
+  useEffect(() => {
+    if (data) {
+      toast.success('수정되었습니다.');
+      void getCategoryRefetch();
+      setCurrentPage(1);
+      onChangeViewMode && onChangeViewMode();
+    } else if (error) {
+      toast.error((error as Error).message);
+    }
+  }, [data, error]);
+
+  return {
+    categoryUpdateRequest,
+  };
 }
 
 export function useDeleteCategoryRequest() {
