+++ src/pages/InfoPage/InfoCategoryManage/index.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react';
 import { Table, Thead, Tbody, Tr, Th, TableCaption, TableContainer } from '@chakra-ui/react';
 import { PlusSquareIcon } from '@chakra-ui/icons';
 import RightContainer from '../../../components/layout/RightContainer';
@@ -6,46 +5,23 @@ import AlertDialogModal from '../../../components/features/AlertDialogModal';
 import CategoryInput from '../../../components/ui/CategoryUI/CategoryInput';
 import Pagination from '../../../components/features/Pagination';
 import AddMode from '../../../components/ui/CategoryUI/AddMode';
-import { useGetCategoryRequest } from '../../../api-hooks/InfoCategory';
-import toast from 'react-hot-toast';
-
-interface CategoryType {
-  id: number;
-  typeName: string;
-  description: string;
-}
+import useInfoCategoryManage from './useInfoCategoryManage';
 
 function InfoCategoryManage() {
-  const [categoryList, setCategoryList] = useState<CategoryType[]>([]);
-  const [isAddClick, setIsAddClick] = useState(false);
-  const [isAlertOpen, setIsAlertOpen] = useState(false);
-  const [location, setLocation] = useState('');
-  const [selectedItemIndex, setSelectedItemIndex] = useState(0);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalPosts, setTotalPosts] = useState(1);
-
-  const { data: categoryResult, isError: categoryError } = useGetCategoryRequest({
-    page: currentPage - 1,
-  });
-
-  useEffect(() => {
-    if (categoryResult) {
-      setCategoryList(categoryResult?.data.articleTypeList);
-      setTotalPosts(categoryResult?.data.totalPage);
-    } else if (categoryError) {
-      toast.error('카테고리 목록을 불러오는데 실패했습니다.');
-    }
-  }, [categoryResult, categoryError]);
-
-  const handleCategoryAddButton = () => {
-    setIsAddClick(true);
-  };
-
-  const handleAlertDialog = (categoryId: number) => {
-    setLocation('카테고리');
-    setSelectedItemIndex(categoryId);
-    setIsAlertOpen(true);
-  };
+  const {
+    isAddClick,
+    setIsAddClick,
+    isAlertOpen,
+    setIsAlertOpen,
+    location,
+    selectedItemIndex,
+    currentPage,
+    setCurrentPage,
+    categoryList,
+    totalPosts,
+    handleCategoryAddButton,
+    handleAlertDialog,
+  } = useInfoCategoryManage();
 
   return (
     <RightContainer title={'정보 게시판 카테고리 관리'}>
@@ -75,7 +51,7 @@ function InfoCategoryManage() {
                 <CategoryInput
                   key={category.id}
                   category={category}
-                  handleAlertDialog={() => handleAlertDialog(category.id)}
+                  handleAlertDialog={() => handleAlertDialog(category.id!)}
                   currentPage={currentPage}
                   setCurrentPage={setCurrentPage}
                 />
